Allow quote in unquoted field instead of aborting
According to RFC 4180, a quote in an unquoted field is not allowed.
(„If fields are not enclosed with double quotes, then double quotes may
not appear inside the fields.“ Section 2, number 5)

However, the default behaviour of Excel and LibreOffice-Calc is to
allow double quotes in an unquoted field when importing csv data.

Example:
abc","def"

Importing this csv with Excel or LibreOffice-Calc and then exporting it
generates this csv output:
"abc""",def